+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/BoundingBoxTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e0dc2d58ac5d0134f93959df8fe4b0fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
