+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/BoundingBoxVisuals.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 52460df9965ac7f48af70ae7eab28f6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
